Cleaned up file system
