+++ src/app/angular/components/analysis/analysis.component.html
