+++ src/app/angular/components/analysis/analysis.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from "@angular/core";
+
+@Component({
+  selector: "app-analysis",
+  templateUrl: "./analysis.component.html",
+  styleUrls: ["./analysis.component.scss"]
+})
+export class AnalysisComponent implements OnInit {
+  constructor() {}
+
+  ngOnInit() {}
+}
