+++ src/app/angular/components/analysis/analysis.module.ts
@@ -1,11 +1,10 @@
 import { SharedModule } from "../shared/shared.module";
-import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { AnalysisComponent } from "./analysis.component";
 
 @NgModule({
   declarations: [AnalysisComponent],
-  imports: [CommonModule, SharedModule],
+  imports: [SharedModule],
   exports: [AnalysisComponent]
 })
 export class AnalysisModule {}
