+++ src/app/angular/components/home/home.component.html
@@ -1,0 +1,1 @@
+<div class="container bg-primary"></div>
