+++ src/app/angular/components/home/home.module.ts
@@ -1,8 +1,6 @@
 import { NgModule } from "@angular/core";
 import { CommonModule } from "@angular/common";
-
 import { HomeRoutingModule } from "./home-routing.module";
-
 import { HomeComponent } from "./home.component";
 import { SharedModule } from "../shared/shared.module";
