+++ src/app/angular/components/shared/navbar/navbar.component.html
@@ -1,4 +1,4 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light mb-3">
+<nav class="navbar navbar-expand-sm navbar-light bg-light mb-3">
   <span class="navbar-brand">
     <img
       src="../../../../assets/ap-desktop.svg"
