+++ src/app/app-routing.module.ts
@@ -1,16 +1,11 @@
 import { NgModule } from "@angular/core";
 import { Routes, RouterModule } from "@angular/router";
-import { PageNotFoundComponent } from "./shared/components";
 
 const routes: Routes = [
   {
     path: "",
     redirectTo: "home",
     pathMatch: "full"
-  },
-  {
-    path: "**",
-    component: PageNotFoundComponent
   }
 ];
