+++ src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
-import { ElectronService } from "./core/services";
 import { TranslateService } from "@ngx-translate/core";
 import { AppConfig } from "../environments/environment";
+import { ElectronService } from "./electron/services/electron/electron.service";
 
 @Component({
   selector: "app-root",
