+++ src/app/app.module.ts
@@ -4,14 +4,14 @@ import { BrowserModule } from "@angular/platform-browser";
 import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { HttpClientModule, HttpClient } from "@angular/common/http";
-import { CoreModule } from "./core/core.module";
-import { SharedModule } from "./shared/shared.module";
 import { AppRoutingModule } from "./app-routing.module";
 import { TranslateModule, TranslateLoader } from "@ngx-translate/core";
 import { TranslateHttpLoader } from "@ngx-translate/http-loader";
-import { HomeModule } from "./home/home.module";
 import { AppComponent } from "./app.component";
-import { AnalysisModule } from "./analysis/analysis.module";
+import { CoreModule } from "./electron/electron.module";
+import { SharedModule } from "./angular/components/shared/shared.module";
+import { HomeModule } from "./angular/components/home/home.module";
+import { AnalysisModule } from "./angular/components/analysis/analysis.module";
 
 // AoT requires an exported function for factories
 export function HttpLoaderFactory(http: HttpClient): TranslateHttpLoader {
