--- src/app/core/services/index.ts
@@ -1,1 +1,0 @@
-export * from "./electron/electron.service";
