+++ src/app/electron/models/analysis.ts
