+++ src/app/electron/models/analysisTypes.ts
