+++ src/app/electron/models/terminal.ts
