--- src/app/home/home.component.html
@@ -1,5 +1,0 @@
-<div class="container">
-  <h1 class="title">
-    {{ 'PAGES.HOME.TITLE' | translate }}
-  </h1>
-</div>
