--- src/app/shared/components/index.ts
@@ -1,2 +1,0 @@
-export * from "./page-not-found/page-not-found.component";
-export * from "./navbar/navbar.component";
