--- src/app/shared/components/page-not-found/page-not-found.component.html
@@ -1,3 +1,0 @@
-<p>
-  page-not-found works!
-</p>
