--- src/app/shared/directives/index.ts
@@ -1,1 +1,0 @@
-export * from "./webview/webview.directive";
