--- src/app/shared/directives/webview/webview.directive.spec.ts
@@ -1,8 +1,0 @@
-import { WebviewDirective } from "./webview.directive";
-
-describe("WebviewDirective", () => {
-  it("should create an instance", () => {
-    const directive = new WebviewDirective();
-    expect(directive).toBeTruthy();
-  });
-});
