--- src/app/shared/directives/webview/webview.directive.ts
@@ -1,8 +1,0 @@
-import { Directive } from "@angular/core";
-
-@Directive({
-  selector: "webview"
-})
-export class WebviewDirective {
-  constructor() {}
-}
